requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LangChain/LangGraph - latest compatible versions
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.50
langgraph-prebuilt>=0.1.0

# Telegram
python-telegram-bot[webhooks]>=22.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
redis>=5.0.0

# Background Tasks
arq>=0.26.0
apscheduler>=3.10.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
